<refactor>: INT-24 Update - Multi-model project Gradle -Добавил 3 модуля Server,Client,Game -Добавил в .gitignore !**/src/ чтобы были все папки тестов
