--- Client/src/main/java/org/example/ClientApplication.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public class ClientApplication {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
