--- Client/src/main/java/org/example/UI/CUI.java
@@ -1,4 +1,0 @@
-package org.example.UI;
-
-public class CUI {
-}
