--- Client/src/main/java/org/example/UI/GUI.java
@@ -1,4 +1,0 @@
-package org.example.UI;
-
-public class GUI {
-}
