--- Game/src/main/java/org/example/GameEntity/Board/Board.java
@@ -1,4 +1,0 @@
-package org.example.GameEntity.Board;
-
-public class Board {
-}
