--- Game/src/main/java/org/example/GameEntity/Units/GeneralBuff.java
@@ -1,4 +1,0 @@
-package org.example.GameEntity.Units;
-
-public interface GeneralBuff {
-}
