--- Server/src/main/java/org/example/Controller/UserRequest.java
@@ -1,4 +1,0 @@
-package org.example.Controller;
-
-public class UserRequest {
-}
