--- Server/src/main/java/org/example/Dto/BoardResponseDto.java
@@ -1,4 +1,0 @@
-package org.example.Dto;
-
-public class BoardResponseDto {
-}
