--- Server/src/main/java/org/example/Dto/MoveRequestDto.java
@@ -1,4 +1,0 @@
-package org.example.Dto;
-
-public class MoveRequestDto {
-}
