--- Server/src/main/java/org/example/ServerApplication.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public class ServerApplication {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
