+++ client/src/main/java/io/deeplay/ClientApplication.java
@@ -1,0 +1,7 @@
+package io.deeplay;
+
+public class ClientApplication {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
