+++ client/src/main/java/io/deeplay/ui/CUI.java
@@ -1,0 +1,4 @@
+package io.deeplay.ui;
+
+public class CUI {
+}
