+++ client/src/main/java/io/deeplay/ui/GUI.java
@@ -1,0 +1,4 @@
+package io.deeplay.ui;
+
+public class GUI {
+}
