+++ game/src/main/java/io/deeplay/Main.java
@@ -1,0 +1,7 @@
+package io.deeplay;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
