+++ game/src/main/java/io/deeplay/game_mechanics/Move.java
@@ -1,0 +1,4 @@
+package io.deeplay.game_mechanics;
+
+public class Move {
+}
