+++ game/src/main/java/io/deeplay/game_mechanics/Placment.java
@@ -1,0 +1,4 @@
+package io.deeplay.game_mechanics;
+
+public class Placment {
+}
