+++ server/src/main/java/io/deeplay/ServerApplication.java
@@ -1,0 +1,7 @@
+package io.deeplay;
+
+public class ServerApplication {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
