+++ server/src/main/java/io/deeplay/controller/UserRequest.java
@@ -1,0 +1,4 @@
+package io.deeplay.controller;
+
+public class UserRequest {
+}
