+++ server/src/main/java/io/deeplay/dto/PlacmentRequestDto.java
@@ -1,0 +1,4 @@
+package io.deeplay.dto;
+
+public class PlacmentRequestDto {
+}
